use HttpApi for playground sharing
