+++ apps/web/src/features/playground/atoms/share.ts
@@ -3,7 +3,7 @@ import * as Effect from "effect/Effect"
 import * as Layer from "effect/Layer"
 import * as AsyncResult from "effect/unstable/reactivity/AsyncResult"
 import * as Atom from "effect/unstable/reactivity/Atom"
-import { ShortenClient } from "@/services/shorten/client"
+import { ShortenClient } from "@/features/playground/services/shorten/client"
 import type { AtomWorkspaceHandle } from "./workspace"
 import { WorkspaceCompression } from "../services/compression"
 import { WorkspaceDownload } from "../services/download"
@@ -38,7 +38,7 @@ export const shareAtom = Atom.family((handle: AtomWorkspaceHandle) => {
         workspace,
         container.readFileString,
       )
-      const hash = yield* client.shorten({ text: compressed })
+      const hash = yield* client.shorten({ payload: { text: compressed } })
       const url = new URL(window.location.pathname, window.location.origin)
       url.hash = hash
