+++ apps/web/src/features/search/api.ts
@@ -1,18 +1,13 @@
-import * as Effect from "effect/Effect"
-import * as Layer from "effect/Layer"
-import { HttpApiBuilder } from "effect/unstable/httpapi"
-import { SearchApi } from "./contract"
-import { Search } from "./service"
+import * as Schema from "effect/Schema"
+import { HttpApi, HttpApiEndpoint, HttpApiGroup } from "effect/unstable/httpapi"
+import { SearchError, SearchResult } from "./domain"
 
-const SearchHandlers = HttpApiBuilder.group(
-  SearchApi,
-  "search",
-  Effect.fn(function* (handlers) {
-    const search = yield* Search
-    return handlers.handle("search", ({ query }) => search.search(query.query))
+export class SearchApiGroup extends HttpApiGroup.make("search").add(
+  HttpApiEndpoint.get("search", "/api/search", {
+    query: { query: Schema.String },
+    success: Schema.Array(SearchResult),
+    error: [SearchError],
   }),
-).pipe(Layer.provide(Search.layer))
+) {}
 
-export const SearchLayer = HttpApiBuilder.layer(SearchApi).pipe(
-  Layer.provide(SearchHandlers),
-)
+export class SearchApi extends HttpApi.make("searchApi").add(SearchApiGroup) {}
