+++ apps/web/src/pages/api/playground/shorten/[hash].ts
@@ -1,0 +1,6 @@
+import type { APIRoute } from "astro"
+import { handler } from "@/features/playground/services/shorten/server"
+
+export const prerender = false
+
+export const GET: APIRoute = ({ request }) => handler(request)
