+++ apps/web/src/pages/api/search.ts
@@ -1,7 +1,7 @@
 import type { APIRoute } from "astro"
 import * as Layer from "effect/Layer"
 import { HttpRouter, HttpServer } from "effect/unstable/http"
-import { SearchLayer } from "@/features/search/api"
+import { SearchLayer } from "@/features/search/server"
 
 export const prerender = false
